requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
Sphinx==1.2.3
beautifulsoup4==4.3.2
caldav==0.2.2
coverage==3.7.1
docutils==0.12
fuzzywuzzy==0.5.0
geopy==1.9.1
icalendar==3.8.4
lxml==3.4.2
nose==1.3.4
python-dateutil==2.4.1
pytz==2014.10
requests==2.5.3
six==1.9.0
vobject==0.6.6
wsgiref==0.1.2
